+++ docs/sync.js
@@ -151,6 +151,15 @@ class App {
   removeUser(user) {}
 
   /**
+   * Deletes the user.
+   * NOTE: This irrecoverably deletes the user from the device as well as the server!
+   *
+   * @param {Realm.User} user - The user to delete.
+   * @returns {Promise<void>}
+   */
+  deleteUser(user) {}
+
+  /**
    * Client for the email/password authentication provider.
    *
    * @example
@@ -655,14 +664,6 @@ class ApiKeyAuth {
   fetchAll() {}
 
   /**
-   * Deletes a user API key associated with the current user.
-   *
-   * @param {string} id - The id of the API key to delete.
-   * @returns {Promise<void>}
-   */
-  delete(id) {}
-
-  /**
    * Enables a user API key associated with the current user.
    *
    * @param {string} id - The id of the API key to enable.
@@ -677,6 +678,14 @@ class ApiKeyAuth {
    * @returns {Promise<void>}
    */
   disable(id) {}
+
+  /**
+   * Deletes a user API key associated with the current user.
+   *
+   * @param {string} id - The id of the API key to delete.
+   * @returns {Promise<void>}
+   */
+  delete(id) {}
 }
 
 /**