+++ packages/realm-web/CHANGELOG.md
@@ -5,7 +5,7 @@
 * None
 
 ### Enhancements
-* None
+* Added `Realm.App#deleteUser()` to delete a sync user from a MongoDB Realm app. ([#4006](https://github.com/realm/realm-js/issues/4006))
 
 ### Fixed
 * None